Refactor representation of APL scalars.
APL scalars were being represented by the APLArray class, when the shape was [] and the data was the value itself. This went against all other APL arrays, for which the data was a Python list of things. This meant that various places of the code had to special case the scalar case, which was becoming annoying. In hindsight, it is perfectly acceptable to have a scalar be distinguished from all other arrays with the empty shape.