Make iota distinguish between scalar and one element vector inputs.
When the argument to (iota) is a vector, (iota) produces a nested array with vectors. This just not happens when the argument is a 1 element vector. Now there is a distinction, and doing applying (iota) to a 1 element vector now produces a vector of 1 element vectors that can be used to index the original vector.
In short, now we have [((iota)4) (match) 0 1 2 3] and [((iota),4) (match) (,0) (,1) (,2) (,3)].